test(common): add main and chart test with applitools and cypress
