+++ .travis.yml
@@ -22,7 +22,8 @@ install:
 script:
   - npm run build:lib
   - npm run test -- --watch=false --no-progress --browsers=ChromeHeadless
-  - ng serve --prod &
+  - ng serve &
   - $(npm bin)/wait-on http-get://localhost:4200/#
-  - npm run test:snapshot
+  - npm run test:main
+  - npm run test:screenshot
   - kill $(jobs -p) || true
