+++ cypress/fixtures/example.json
@@ -1,5 +1,3 @@
 {
-  "name": "Using fixtures to represent data",
-  "email": "hello@cypress.io",
-  "body": "Fixtures are a great way to mock data for responses to routes"
+  "componentsUrls": ["#LineChart", "#BarChart", "#DoughnutChart", "#RadarChart", "#PieChart", "#PolarAreaChart", "#BubbleChart", "#ScatterChart", "#DynamicChart"]
 }
